Documentation and refinement on `FragmentRouterBuilder`
