+++ kompass-android/src/main/java/io/sellmair/kompass/android/fragment/PopRetainRootImmediateSyntax.kt
@@ -8,6 +8,12 @@ interface PopRetainRootImmediateSyntax {
 
     class NotAttachedException(message: String?) : KompassException(message)
 
+
+    /**
+     * Will pop a single element from the routing stack, but will never pop the `root` (the last remaining route)
+     * @return true: If the event was consumed an a route was popped
+     * false: If routing stack has equal or less than one element left and no route was popped
+     */
     fun <T : Route> FragmentRouter<T>.popRetainRootImmediate(): Boolean {
         var didPop: Boolean? = null
 
@@ -28,6 +34,6 @@ interface PopRetainRootImmediateSyntax {
 
 private fun throwNotAttachedException(): Nothing {
     throw PopRetainRootImmediateSyntax.NotAttachedException(
-        "popRetainRootImmediate can only be called between `onResume` and `onPause` and after router.setup"
+        "popRetainRootImmediate can only be called in the correct lifecycle state and after router.setup"
     )
 }
