+++ kompass-android/src/main/java/io/sellmair/kompass/android/fragment/dsl/FragmentRouterBuilder.kt
@@ -1,10 +1,12 @@
 package io.sellmair.kompass.android.fragment.dsl
 
 import android.os.Parcelable
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Lifecycle
 import io.sellmair.kompass.android.fragment.*
-import io.sellmair.kompass.android.fragment.internal.*
+import io.sellmair.kompass.android.fragment.internal.EmptyFragmentTransition
+import io.sellmair.kompass.android.fragment.internal.FragmentContainerLifecycle
+import io.sellmair.kompass.android.fragment.internal.GenericFragmentContainerLifecycle
+import io.sellmair.kompass.android.fragment.internal.createUnsafe
 import io.sellmair.kompass.core.EmptyRouterInstruction
 import io.sellmair.kompass.core.Route
 import io.sellmair.kompass.core.RouterInstruction
@@ -12,7 +14,7 @@ import io.sellmair.kompass.core.plus
 import kotlin.reflect.KClass
 
 @FragmentRouterDsl
-class FragmentRouterBuilder<T : Route>(type: KClass<T>) {
+class FragmentRouterBuilder<T : Route>(private val type: KClass<T>) {
 
     private val typeIsParcelable = Parcelable::class.java.isAssignableFrom(type.java)
 
@@ -30,7 +32,7 @@ class FragmentRouterBuilder<T : Route>(type: KClass<T>) {
     }
 
     private var fragmentRoutingStackBundleSyntax: FragmentRoutingStackBundleSyntax<T>? = when {
-        typeIsParcelable -> ParcelableFragmentRoutingStackBundler.createUnsafe()
+        typeIsParcelable -> ParcelableFragmentRoutingStackBundleSyntax.createUnsafe()
         else -> null
     }
 
@@ -40,18 +42,46 @@ class FragmentRouterBuilder<T : Route>(type: KClass<T>) {
             detachEvent = Lifecycle.Event.ON_PAUSE
         )
 
-
+    /**
+     * Specify a custom [FragmentStackPatcher]
+     * @see FragmentStackPatcher
+     * @see DefaultFragmentStackPatcher
+     */
     @FragmentRouterDsl
     fun fragmentStackPatcher(patcher: FragmentStackPatcher) {
         this.fragmentStackPatcher = patcher
     }
 
+    /**
+     * Specify a custom [FragmentRouteStorageSyntax].
+     *
+     * [ParcelableFragmentRouteStorageSyntax] will be used as default if the base route type [T] is [Parcelable]
+     *
+     * @see FragmentRouteStorageSyntax
+     * @see ParcelableFragmentRouteStorageSyntax
+     */
     @FragmentRouterDsl
     fun fragmentRouteStorage(storage: FragmentRouteStorageSyntax<T>) {
         this.fragmentRouteStorageSyntax = storage
     }
 
+    /**
+     * Specify a custom [fragmentRoutingStackBundleSyntax]
+     *
+     * [ParcelableFragmentRoutingStackBundleSyntax] will be used as default if the base route type [T] is [Parcelable]
+     *
+     * @see FragmentRoutingStackBundleSyntax
+     * @see ParcelableFragmentRouteStorageSyntax
+     */
+    @FragmentRouterDsl
+    fun fragmentRoutingStackBundler(bundler: FragmentRoutingStackBundleSyntax<T>) {
+        this.fragmentRoutingStackBundleSyntax = bundler
+    }
 
+    /**
+     * Configures a [FragmentMap] that will be used by the router.
+     * Can be invoked multiple times
+     */
     @FragmentRouterDsl
     fun routing(init: FragmentMapBuilder<T>.() -> Unit) {
         this.fragmentMap += FragmentMapBuilder<T>().also(init).build()
@@ -84,39 +114,34 @@ class FragmentRouterBuilder<T : Route>(type: KClass<T>) {
 
     private fun requireFragmentRouteStorage(): FragmentRouteStorageSyntax<T> {
         return fragmentRouteStorageSyntax ?: throw KompassFragmentDslException(
-            ""
+            """
+                Missing `FragmentRouteStorageSyntax`!
+                Either specify one with
+
+                    FragmentRouter {
+                        ...
+                        fragmentRouteStorage(MyFragmentRouteStorageSyntax())
+                    }
+
+                Or let ${type.java.simpleName} implement `Parcelable` to use the default implementation
+            """.trimIndent()
         )
     }
 
 
     private fun requireFragmentRoutingStackBundler(): FragmentRoutingStackBundleSyntax<T> {
         return fragmentRoutingStackBundleSyntax ?: throw KompassFragmentDslException(
-            ""
+            """
+                Missing `FragmentRoutingStackBundleSyntax`!
+                Either specify one with
+
+                    FragmentRouter {
+                        ...
+                        fragmentRoutingStackBundler(MyFragmentRoutingStackBundleSyntax())
+                    }
+
+                Or let ${type.java.simpleName} implement `Parcelable` to use the default implementation
+            """.trimIndent()
         )
     }
 }
-
-
-class FragmentTransitionBuilder {
-
-    var transition: FragmentTransition = EmptyFragmentTransition
-
-    @FragmentRouterDsl
-    fun register(transition: FragmentTransition) {
-        this.transition += transition
-    }
-
-    @JvmName("registerGeneric")
-    @FragmentRouterDsl
-    inline fun <reified ExitFragment : Fragment, reified ExitRoute : Route,
-            reified EnterFragment : Fragment, reified EnterRoute : Route> register(
-        transition: GenericFragmentTransition<ExitFragment, ExitRoute, EnterFragment, EnterRoute>
-    ) = register(transition.reified().erased())
-
-
-    internal fun build(): FragmentTransition {
-        return transition
-    }
-}
-
-
